Place smaller areas atop larger ones
If we don’t assign any z-index to the areas, then they overlap each
other in the order in which they appear in the HTML document, which can
make it hard to target certain areas, depending on which order the
corresponding statements appear in on the item. Solving this in general
– finding some ideal ordering of the areas that minimizes overlap –
sounds like a hellishly hard problem to solve, but we can approximate it
by making the z-index inversely proportional to the size of the area.
This places smaller areas atop larger ones, solving the problem of
larger areas completely covering smaller ones.