Add proper index page
The index page briefly explains what the tool does and offers you a form
for the /item/ endpoint. The “by item and property” heading is there
because I anticipate the addition of a second endpoint (and index page
form) to find items by their “relative position within image” qualifier
value, so that this endpoint of the tool can be used as the formatter
URL for that property (in theory, the value can be ambiguous, but I
would expect collisions to be extremely rare).