Limit image size
Some images have *very* high resolution, so let’s make sure they don’t
overflow their container. Limiting the image width is sufficient for
this: if the img is not too wide, then the .image will not be too wide
either, and therefore all the .depicteds (which are positioned against
the .image) will still be positioned correctly.

Also limits the image’s height, because that makes sense in general, but
since the page height is not limited, this doesn’t currently have any
effect (though that may change later).