Make active area more visible
In images with lots of depicted elements, it can be hard to determine
which lines belong to which areas. Make this clearer by giving the
currently hovered area a thicker line.

This also changes the lines from borders to outlines. Borders occupy
space on the element, so making them thicker takes away some space from
the text, potentially causing it to reflow, which looks ugly.  Outlines,
which lie outside the element’s content, don’t have this problem.