Add endpoint to find images by IIF region string
This endpoint is intended to be used for the formatter URL of the
“relative position within image” property. Usually, the value should be
unique and only find a single item (the one where the link was clicked),
but in case of multiple items we show them all.

Most of the code for this is reused from item_and_property() and
therefore extracted into a separate function, load_item_and_property(),
which takes care of parsing the data, loading item labels and so on.
Unfortunately, its error handling currently doesn’t make sense anymore –
I’ll fix that later.