Fix JSON response loading
json.load [1] only accepts binary files, such as HTTPResponse objects,
since Python 3.6. As Toolforge only has Python 3.4, we have to go for
the slightly more inefficient route of loading the whole response into
memory, decoding it in-memory, and then parsing it in-memory. Luckily,
none of the responses are expected to be large enough for this to matter
very much.

[1]: https://docs.python.org/3/library/json.html#json.load