Demo of wd image positions as IIIF annotations
This is for demo purposes only, and could be made richer.
It adds two new routes to the tool-wd-image-positions application:

- A IIIF Manifest, to provide a canvas (annotation space) for the image
- An annotation list, that contains the annotations that target the canvas

Examples:

http://localhost:5000/iiif/Q1231009/manifest.json
http://localhost:5000/iiif/Q1231009/list/annotations.json

This can be made more robust with some error handling. The annotations
are also very sparse, I made them as simple as possible just to
demonstrate them working in Mirador:

https://www.dropbox.com/s/782rc0kadlhq81b/napoleon-mirador.jpg?dl=0

The Preview in Mirador will launch a version of Mirador for the
generated manifest; you'll need to then turn on the annotation layer
(the speech bubble icon) to see the annotations.