Split image_attribution into params/response functions
This makes it possible to combine this API request with others, by
calling several functions to add params, making one request, and then
calling several functions to process the response again. The original
function is retained as a wrapper, in case you don’t want to combine any
requests, and for now we keep using that function, though future commits
will also get the image URL in the same request.