Handle OAuth identify() failures
I’m soon going to update the OAuth consumer version so that other wikis
than Wikidata can be edited, which will invalidate old sessions. Treat
any error in identify() as an indication that the access token is
invalid and should be removed from the session, and the user treated as
not logged in.