Add alternative API for adding qualifiers
This new API allows specifying the domain (though only Wikidata is
allowed so far), and moves the other parameters from the URL to the POST
data, since I think that’s the more appropriate place for them. (Its URL
is also versioned.)

Because this may result in edits on other wikis, the edit summary used
by the new API links to the Wikidata page for the tool using the d:
prefix, whereas this is not necessary for the old API.

The old API is kept for now, though it should be removed soon.