Fix overflowing text enlarging depicteds’ outlines
If the content of an element overflows its block, the outline will
reflect the dimensions of the content, not of the block, giving the
impression of a larger depicted area than is correct. Specify overflow:
none to fix this, and add the label and item ID as the title attribute
to compensate for the labels not being fully visible in some cases.

For the rationale of using outline rather than border in this tool (with
border, we wouldn’t have this problem), see commit 4c7f8fa7bd.