Add support for showing somevalue/novalue statements
Instead of ignoring “unknown value” and “no value” statements when
loading the depicted items, we treat them like depicteds without an item
ID, and with the translation of the “some value” / “no value” message as
the label. Other parts of the app are updated to deal with the fact that
a depicted may no longer have an item ID – for the image page, we
actually show them, but for the IIIF template we just skip them.

There’s no editing support yet, that’ll come later.

Note that this introduces a new dependency (requirements.txt), so
deployment must include a pip install.