Support editing somevalue/novalue when logged in
Editing somevalue and novalue statements in logged in mode is pretty
straightforward, because all of our editing is based on the statement ID
anyways, not on the value. In QuickStatements mode, that wouldn’t work
so well, so we don’t support it there – there’s no way to tell multiple
somevalue statements apart. (In theory, the same is true if you have
multiple statements with the same item value, but I hope that should be
rare, and it’s not a problem I want to deal with right now.)