Add support for adding more “depicted” statements
This adds a form allowing users to add more “depicted” statements to the
image, for which they can then subsequently define regions. (This will
result in two separate edits; this is a lot easier to implement and
seems acceptable to me.) Technically, “some value” and “no value” are
fully supported, but both corresponding buttons are currently hidden –
the “no value” one most likely permanently, because I see no real value
in it, but the “some value” will hopefully be enabled once the rest of
the Structured Data on Commons ecosystem is ready for it (most
importantly, the WikibaseMediaInfo UI should allow you to edit these
values). There’s no real reason to disable it on Wikidata, but no strong
need for it there either, so disabling it everywhere is simpler.

All of this is only available in logged in mode, though, not via
QuickStatements.