Don’t bother using hyphenated language codes
They almost never match between MediaWiki and BCP 47 (mainly due to the
MediaWiki version not capitalizing the country/region or alphabet
parts), and the API raises a warning in that case, which clogs up the
uwsgi.log. We already added their base version (before the hyphen); now,
use only that. (And add a comment indicating that this could be made
more accurate using the languageinfo API, though we’d want to cache it
to avoid slowing down every request.)