Allow editing regions in user interface
If the user is logged in (there’s no way to target an existing qualifier
in QuickStatements), add an “edit a region” button. When it is clicked,
clicking one of the regions will select it for editing instead of
opening the link, and add a cropper using that region’s coordinates.
After adjusting the cropper as needed, clicking the button again will
send an edit for that region.

The actual saving functionality is extracted into a function, since most
of the code is shared between the “new region” and “edit region” cases.
The edge cases get a bit finicky, especially around the interaction with
QuickStatements mode; honestly, I’ll probably remove that mode soon, I
don’t expect it to be used very much anymore.