Don’t use mwoauth.identify() to get user name
This is a weird interface with an undocumented response format, and it’s
completely redundant with the standard userinfo API, so let’s use that
instead. Error handling changes a bit in that case. (I tested that the
error handling still works, using the credentials of the 1.1 consumer
that I still had lying around.)

Also, restructure the code to avoid duplicating the login markup.