Make SPARQL queries with proper user agent
Previously, we were using the default Python urllib user agent (which
looks like Python-urllib/3.10 according to some local testing). Ouch.

We could also use the SPARQLWrapper library, but I don’t really see a
reason to, to be honest… the main feature it could offer would be to
escape the iiif_region for us, and it doesn’t seem to have anything like
that, as far as I can tell.