Support adding “named place on map” statements
Now the JS interface can also add new statements with that property.
(The QuickStatements code still has a hard-coded P180, but I want to
remove that soon anyways.) This requires an extra label in the
depicted_properties.

In the OOUI code, I couldn’t find a way to get from the propertyIdInput
variable to the actual value – it looks like the relevant element is
cloned somewhere, for some reason. Work around that with a global ID.

In the server-side code, the property_id in the form data is optional
for the moment (defaulting to P180), in case people still have tabs open
with the old JS code. Could be removed later, though.

Part of #2.