Fix item ID input handling
Apparently, the EntityInputWidget can return the wrong entity ID from
getData().getSerialization() (the ID of the first search result, even if
the user selected another one, but not always). In SDC, this is
presumably not an issue because as soon as an item ID is selected, it is
used for preparing a new statement (not yet saved), and the input itself
is cleared for the next potential statement. Do something similar here,
and only trust item IDs we get out of the “change” event. (“add” would
also work, but is less convenient because it has an extra argument.)