Add info message when users cannot edit
While this tool initially started as a *viewer* of “relative position
within image” qualifiers, it quickly became an *editor* of this data,
and this may in fact be its primary use now. As such, we should let
users know when they can’t edit, whether that’s because they’re not
logged in and/or because JavaScript is disabled.

The notice is styled like an alert, but doesn’t have role="alert", since
it’s not information that requires the user’s immediate attention (if
users are not interested in editing, they might be okay with seeing this
message on every page view) – I just like Bootstrap’s alert styles
better than the card styles (which would be one alternative).

Note that authentication_area() in app.py takes care to not add a
useless login link when the app doesn’t have OAuth configured, but we
don’t repeat this complexity in the edit_info(); I’m just not very
interested in local development of this tool without OAuth anymore.