Scroll only image when scaling up
Put the image inside a wrapper, and scroll only that wrapper if the
image overflows it, rather than scrolling the whole page. This keeps the
rest of the user interface more accessible.

The CSS needs to be tweaked a little due to the extra element; to make
the hidden buttons work, the wrapper needs to be tagged with the
--active class as well (though we still keep it on the image too).