Make image height less than full viewport height
Leave a bit of space around the image, so that parts of the user
interface can be visible at the same time: roughly enough space for the
scale slider above and a few depicteds without region below. But on very
small screens, make sure this doesn’t make the max-height negative.