Refactor Esc handling JS code
I want to add “cancel” buttons, so the code apart from checking whether
the keypress was Escape or not will be needed for that too. Extract the
onEscape() handler so we can keep it out of the handler functions.

Also, this made me realize that we’re generally failing to remove the
event handler after successfully editing; fix that.