Move from Diffusion to Wikimedia GitLab
Apparently Wikimedia GitLab is considered stable enough for personal
projects like this by now, and getting rid of Diffusion is always a good
thing. Note that we stay on GitHub for CI purposes, at least until
global job runners are available (T297426).

The list of contribution methods in the README can perhaps be removed or
at least abridged later; the main point of the old list was
“Differential is terrible and I prefer literally anything else”, and
with Diffusion gone, that’s no longer a concern, so it’s probably enough
to just let people choose between pull requests on either platform for
themselves (or use the email workflow if they really want to).