Refactor colors in depicted.css
Someone suggested to use a semi-transparent background instead of an
outline (I’m not sure if they’d want to be named here or not ^^). I’m
not convinced by it, but this changed organization of the CSS (with no
effect) at least makes it easier to try out. Users can add the
background, and potentially remove the outline, in their user CSS, if
they want. (That includes browser extensions like Stylish, or also, for
the user script [1], their common.css on Wikimedia Commons.)

[1]: https://www.wikidata.org/wiki/User:Lucas_Werkmeister/Wikidata_Image_Positions#User_script