Disable underline for depicted links
Bootstrap 5 underlines links by default even if not hovered; for the
depicted links, this is too much, since the underline can obscure other
depicteds or the underlying image. Disable the underline unless the link
is hovered or focused; MediaWiki core’s elements.less (shared by most if
not all skins, I assume) uses the same pseudo-classes.

Put this style in image.css, not depicted.css, as it doesn’t make sense
on-wiki; but add pointers between the two files.