URL-encode image title for srcset
This is needed for files with question marks in the title, e.g.
[[File:CSD Berlin 2022 - Lucas Werkmeister - 49 - Do You Think You’re
More Tired Of The War Than We Are?.jpg]] [1]. Compare also the
corresponding fix to the user script [2].

Note that we don’t need to URL-encode anything for the src attribute: we
got that full URL from the API, MediaWiki is responsible for encoding it
as necessary.

[1]: https://commons.wikimedia.org/wiki/File:CSD_Berlin_2022_-_Lucas_Werkmeister_-_49_-_Do_You_Think_You%E2%80%99re_More_Tired_Of_The_War_Than_We_Are%3F.jpg
[2]: https://commons.wikimedia.org/w/index.php?title=User:Lucas_Werkmeister/wd-image-positions.js&diff=677348415&oldid=653031686