Only load edit interface when logged in
This greatly reduces the amount of data loaded for visitors who aren’t
logged in. Additionally, even for users who are logged in, it’s good if
image.js loads faster and doesn’t require importing other libraries, so
that it can replace the “image scale” interface as quickly as possible.