Disable autoCrop / initial area selection
In my experience, the initial crop suggested by Cropper.js is never
useful, but since it covers a lot of the image, I often have to
“discard” it by selecting a random small area at the margin first, then
drawing the actual selection I want. Let’s avoid that by just not having
an autocrop. If users click the “use this region” button before they
made a selection, show them an error and don’t save an edit. (alert()
isn’t very nice, but I can’t be bothered to look for the right event to
determine whether the user started making a selection, which would let
us change the button label appropriately.)

If anyone complains about this, an alternative approach would be to
change autoCropArea from its default (0.8, 80% of the image) to
something smaller, so that the initial area is smaller and covers less
of the image. But let’s try disabling autoCrop first.