Add toolforge_i18n, make index.html translatable
This introduces toolforge_i18n as a Git submodule (it will eventually
become a proper Python package, but not yet), and adds messages for
everything visible in index.html (including base.html). The rest of the
tool will be made translatable separately.

Includes a few derived messages, but I also just straight-up reused a
few messages; if the “derivation” is just the string identity function,
then it feels a bit pointless to introduce a derived message – if I ever
need to split the message, splitting its uses in the code shouldn’t be
the hard part anyways, so preempting such a split via derived messages
doesn’t bring a huge benefit, I think. I see some sense in it for the
“tool name” messages, since they’re used quite differently, but for the
others I’m happy to just reuse the message key.

The presence of both `message` and `messages` in app.py now (which are
very different) is unfortunate, but I can’t be bothered to fix it right
now. I’m guessing the messages module should be renamed.