Make “image scale” translatable
And add the translations to a new data attribute where the JS can load
them from.

In principle, we could also restructure the HTML to put the server-side
label in a separate element (now that modern browsers have :has() in
CSS, we’re a bit more free in how we arrange the DOM for the no-JS
version of the scale inputs), and have the JS read the text content
from there, but that’s more effort than I want to put into it right now,
and anyway some of the other JS parts will need server-sent translations
anyway (edit mode).