Map zh-hant to zh_Hant for Babel
This makes me think toolforge_i18n’s default language_code_to_babel
should be a bit smarter than it currently is – at least ship some
standard replacements for common (or, more to the point, commonly
translated) language codes that need a mapping. But for now this is good
enough to load the zh-hant translations.

Bug: T363626