Add GitLab CI
Since l10n-bot merge requests are sent on GitLab, not on GitHub, we also
want to run the tests there – these tests are very important, checking
that the tool doesn’t crash and even that the messages only contain safe
HTML. (The GitHub actions could in principle be removed now, but for now
I feel like keeping it.)

The venv seems to be required – without it, `pip` is not available
inside the image. I don’t understand it, but I’m fine with it, and
installing the dependencies in a venv probably makes sense anyway.

(Other tools can probably omit the GIT_SUBMODULE_STRATEGY, and once I
turn toolforge_i18n into a proper library installed via pip I’ll be able
to remove it again here as well.)