Make “… with no region specified” translatable
Use the property labels from Wikidata for the interactive dropdown /
select labels, and dedicated messages for “… with no region specified”
per property ID (only two properties at the moment).

Remove the part of image.html that tried to render depicteds without a
well-known property. This was always dead code (we only put the
well-known properties into the depicteds to begin with, so this was only
a guard against me forgetting something somewhere or letting the code
get out of sync, I think), and it’s not really clear what it should look
like now; it would be possible to implement it in some way anyway but I
don’t think it’s worth it.

Bug: T363626