Fix TIFF workaround
This broke because the srcset property now contains a huge amount of
whitespace, so the “adjusted” src was just the empty string. I don’t
think the way we assemble the srcset attribute in the HTML changed since
the workaround was introduced in e7fd7f5c82, so I can only assume that
this used to work because Firefox (and other browsers?) previously
returned a “formatted” / “cleaned up” version of the srcset attribute in
the HTMLImageElement srcset property, whereas now they return the
original attribute contents with whatever whitespace they contain; but I
wasn’t able to find a Bugzilla entry confirming this. But anyway, this
seems to work now.