Fix language fallback in edit interface
The messages are supposed to be plain text, but the translations can
still be HTML if language fallback happened:

    <span lang="en" dir="ltr">Edit a region</span>

So to avoid showing that markup to the user, we should assign the
translations to the innerHTML, not to the textContent. We’ll just have
to rely on the translations tests to make sure that the translations are
safe and don’t contain any unexpected or malicious tag or attribute
names.