Search for items in interface language
Note that we need two “language” parameters: `language` sets the
language in which the search is done, `uselang` sets the language in
which search results are shown. (If you only set language=de, for
instance, searching for „Baum“ will show Q10884 as “tree”.)

The data-uselang attribute isn’t used yet, but included for consistency
with Wikidata Lexeme Forms; IMHO it doesn’t hurt to include it, and it
might be useful later.