Fix logout
We need to clear the access token (set in /oauth/callback), not the
request token (set in /login and already popped in /oauth/callback). I’m
not sure how I got this wrong in 78a1d19531 – it’s correct in
cookiecutter-toolforge and all my other tools, as far as I can tell
(though some of them clear the whole session instead of just popping a
single value from it).