Upgrade to toolforge_i18n 0.0.2
This registers the translations tests in a much nicer way (and enables
assertion rewriting for them too). The translations can’t be imported
anymore, but we can get them as a fixture instead; and the message_key
parameter has to be renamed to avoid collision with the toolforge_i18n
provided fixture (we don’t want to parametrize this test with *all*
message keys).