Split dev requirements from prod requirements
This shrinks the production venv a little bit, though not by much
(134 MiB → 120 MiB; this tool doesn’t use mypy, and the biggest package
in the venv is babel with all its locale-data).

In the GitLab CI, move to the general `python` container image for
consistency with other tools; also, we’re about to move to Python 3.13,
and there’s no toolforge-python313-* image yet anyway.